+++ app/src/main/java/com/iyanuoluwa/imovie/ui/details/DetailsActivity.kt
@@ -10,8 +10,8 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import com.iyanuoluwa.imovie.R
 import com.iyanuoluwa.imovie.data.model.Credits
-import com.iyanuoluwa.imovie.data.remote.ApiCredits
-import com.iyanuoluwa.imovie.ui.main.MainActivity
+import com.iyanuoluwa.imovie.data.remote.MovieApi
+import com.iyanuoluwa.imovie.util.BASE_URL
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -65,10 +65,10 @@ class DetailsActivity : AppCompatActivity() {
     private fun getCredits() {
         val id = intent.getIntExtra("id", 460465)
         val api = Retrofit.Builder()
-            .baseUrl(MainActivity.BASE_URL)
+            .baseUrl(BASE_URL)
             .addConverterFactory(GsonConverterFactory.create())
             .build()
-            .create(ApiCredits::class.java)
+            .create(MovieApi::class.java)
 
         val data = api.getMovieCredits(id)
         data.enqueue(object : Callback<Credits?> {
