+++ app/src/main/java/com/iyanuoluwa/imovie/util/Constants.kt
@@ -1,0 +1,5 @@
+package com.iyanuoluwa.imovie.util
+
+// https://api.themoviedb.org/3/movie/now_playing?api_key=68f286331e8795bd4addf043c1e8423d&language=en-US
+const val BASE_URL = "https://api.themoviedb.org/3/"
+const val API_KEY = "68f286331e8795bd4addf043c1e8423d"
